Rename to GitHubApp
